% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSYNC_package.R
\name{bayesSYNC-package}
\alias{bayesSYNC-package}
\title{bayesSYNC: an R package for high-dimensional functional factor analysis of shared latent dynamics}
\description{
High-dimensional factor model based on functional principal component
analysis (FPCA) using variational inference.
}
\details{
"_PACKAGE"
}
