% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSYNC.R
\name{bayesSYNC_model_choice}
\alias{bayesSYNC_model_choice}
\title{This function performs functional analysis to uncover shared latent dynamics
using Bayesian methods. with model choice procedure for
selection of either the number of FPCA components, L, or the number of latent
factors, Q.}
\usage{
bayesSYNC_model_choice(
  time_obs,
  Y,
  model_choice,
  list_L,
  list_Q,
  K,
  list_hyper = NULL,
  n_g = 1000,
  time_g = NULL,
  tol_abs = 0.001,
  tol_rel = 1e-05,
  maxit = 1000,
  n_cpus = 1,
  verbose = TRUE,
  seed = NULL,
  bool_scale = TRUE
)
}
\arguments{
\item{time_obs}{List of vectors or list of lists of vectors containing time
of observations for univariate or multivariate curves,
respectively.}

\item{Y}{List of vectors (for univariate curves) or list of lists of vectors
(for multivariate curves) with the functional data.}

\item{model_choice}{String indicating whether the model choice is to be
performed for the number of latent factors ("Q") or for the number of
components ("L").}

\item{list_L}{List containing specifications for the choice of the number of
components 'L'. If \code{model_choice = "L"}, then must consist
of two parameters, 'lambda_L' and 'L_max', specifying the
Poisson(lambda_L) prior, truncated to {1, ..., L_max}, on the
number of components L (e.g., lambda_L = 3, L_max = 10);
if \code{model_choice = "K"}, then must consist of an integer
named 'L' representing the (maximum) number of components to be
used (e.g., L = 10).}

\item{list_Q}{List containing specifications for the choice of the number of
factors 'Q'. If \code{model_choice = "Q"}, then must consist of
two parameters, 'lambda_Q' and 'Q_max', specifying the
Poisson(lambda_Q) prior, truncated to {2, ..., Q_max}, on the
number of factors Q (e.g., lambda_Q = 1, L_max = 6);
if \code{model_choice = "L"}, then must consist of an integer
named 'Q' representing the (maximum) number of factors to be
used (e.g., Q = 3).}

\item{K}{Number of O'Sullivan spline functions to be used to represent the
latent functions. If set to \code{NULL} will be set according to the
rule of Ruppert (2002), also enforcing K >=7.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL} default hyperparameters will
be used.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{tol_abs}{Tolerance on the absolute changes in the ELBO.}

\item{tol_rel}{Tolerance on the relative changes in the ELBO.}

\item{maxit}{Number of iterations for the mean-field variational Bayes algorithm. Default maxit = 1000.}

\item{n_cpus}{User-specified number of CPU to be used for parallel execution. Default n_cpus = 1 for serial execution.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run.}

\item{seed}{User-specified seed for reproducibilty.}

\item{bool_scale}{Whether to standardise the variables. The per-variable
within-individual means are first computed, their mean and sd across
individuals are obtained, and used to scale the measurements. Default is TRUE.}
}
\value{
An object containing the resulting MFVB estimates.
}
\description{
This function performs functional analysis to uncover shared latent dynamics
using Bayesian methods. with model choice procedure for
selection of either the number of FPCA components, L, or the number of latent
factors, Q.
}
\references{
Nolan, T. H., Goldsmith, J., & Ruppert, D. (2023). Bayesian Functional
Principal Components Analysis via Variational Message Passing with Multilevel
Extensions. Bayesian Analysis, 1(1), 1-27.

Ruppert, D. (2002). Selecting the number of knots for penalized splines.
Journal of computational and graphical statistics, 11(4), 735-757.
}
\seealso{
\code{\link{run_mfvb_fpca}},
          \code{\link{run_vmp_fpca}}, \code{\link{run_vmp_fpca_model_choice}},
          \code{\link{set_hyper}}, \code{\link{flip_sign}}, \code{\link{display_fit}},
          \code{\link{display_fit_list}}, \code{\link{display_scores}},
          \code{\link{display_eigenfunctions}}
}
