% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flip_sign}
\alias{flip_sign}
\title{Flip the sign of eigenfunctions and corresponding scores}
\usage{
flip_sign(vec_flip, list_Psi_hat, Zeta_hat, zeta_ellipse = NULL)
}
\arguments{
\item{vec_flip}{Vector of size L whose entry l is -1 if sign needs to be
flipped for eigenfunction l = 1, ..., L, or 1 if sign is keep
the same.}

\item{list_Psi_hat}{List or matrix of eigenfunctions as returned by
\code{\link{bayesSYNC}}.}

\item{Zeta_hat}{Matrix of estimated scores as returned by
\code{\link{bayesSYNC}}.}

\item{zeta_ellipse}{95\% posterior credible boundaries for the scores as
return by \code{\link{bayesSYNC}}.}
}
\value{
An object containing the eigenfunctions, scores and credible
        boundaries with altered sign.
}
\description{
This function is used to alter the sign of specified eigenfunctions and
scores. Choosing one eigenfunction over its opposite sign has no effect on
the resulting fits, although one choice of sign may provide more natural
interpretation of the eigenfunction.
}
