% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_grid_objects}
\alias{get_grid_objects}
\title{Create a temporal-grid objects for use in FPCA algorithms.}
\usage{
get_grid_objects(
  time_obs,
  K,
  n_g = 1000,
  time_g = NULL,
  int_knots = NULL,
  format_univ = FALSE
)
}
\arguments{
\item{time_obs}{Vector or list of vectors containing time of observations for
univariate or multivariate curves, respectively.}

\item{K}{Number of O'Sulivan spline functions to be used in the FPCA
algorithms. If set to \code{NULL} will be set according to the rule
of Ruppert (2002), also enforcing K >=7.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{int_knots}{Position of interior knots. Default is \code{NULL} for
evenly placed.}

\item{format_univ}{Boolean indicating whether the univariate format is used
in case of univariate curves.}
}
\value{
C Design matrix C(t) constructed from the set of K spline functions
          based on the observation time grid.

n_g Number of time points in the dense grid.

time_g Dense time grid constructed.

C_g Design matrix C(t) constructed from the set of K spline functions
            based on the dense time grid.
}
\description{
This function is used to create a time grid of desired density and design
matrices based on an observed time grid object.
}
