% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper.R
\name{set_hyper}
\alias{set_hyper}
\title{Gather model hyperparameters.}
\usage{
set_hyper(
  sigma_beta = 1e+05,
  A = 1e+05,
  c_0 = 1,
  d_0 = 1,
  alpha_0 = 1,
  beta_0 = 0.5
)
}
\arguments{
\item{sigma_beta}{Vector of size 1 or 2 for the standard deviation of the
spline coefficients for the mean function.}

\item{A}{Positive real number for the top-level hyperparameter of the
iterated inverse chi-square distributions.}

\item{c_0}{to add}

\item{d_0}{to add}

\item{alpha_0}{to add}

\item{beta_0}{to add}
}
\value{
An object containing the hyperparameter settings to be supplied to
        \code{\link{bayesSYNC}}.
}
\description{
This function is used to construct an object with hyperparameter values for
the FPCA or mFPCA model in a format that can then be supplied to
\code{\link{bayesSYNC}}.
}
\details{
The \code{\link{bayesSYNC}} function can
also be used with default hyperparameter values (i.e., without using
\code{\link{set_hyper}}) by setting their argument \code{list_hyper} to
\code{NULL}.
}
