% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trapint}
\alias{trapint}
\title{Function performing trapesoidal integration.}
\usage{
trapint(xgrid, fgrid)
}
\arguments{
\item{xgrid}{Grid.}

\item{fgrid}{Function on the grid.}
}
\value{
Integration result.
}
\description{
Function performing trapesoidal integration.
}
