% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSYNC.R
\name{bayesSYNC}
\alias{bayesSYNC}
\title{BAYESian model for high-dimensional functional factor analysis of Shared latent dYNamiCs (bayesSYNC)}
\usage{
bayesSYNC(
  time_obs,
  Y,
  L,
  Q,
  K = NULL,
  anneal = c(1, 1.9, 100),
  list_hyper = NULL,
  n_g = 1000,
  time_g = NULL,
  tol_abs = 0.001,
  tol_rel = 1e-05,
  maxit = 1000,
  n_cpus = 1,
  verbose = TRUE,
  seed = NULL,
  bool_scale = TRUE,
  bool_var_spec_prob = FALSE,
  show_factor_ppi_progress = FALSE
)
}
\arguments{
\item{time_obs}{List of vectors or list of lists of vectors containing time
of observations for univariate or multivariate curves, respectively.}

\item{Y}{List of lists of vectors
(for multivariate curves) with the functional data.}

\item{L}{Number of latent dimensions.}

\item{Q}{Number of latent factors.}

\item{K}{Number of O'Sullivan spline functions to be used to represent the
latent functions. If set to \code{NULL} will be set according to the
rule of Ruppert (2002), also enforcing K >=7.}

\item{anneal}{Parameters for annealing scheme. Must be a vector whose first
entry is the type of schedule: 1 = geometric spacing (default),
2 = harmonic spacing or 3 = linear spacing, the second entry is the initial
temperature which must lie in the interval (0, 2) (default is 1.9), and the
third entry is the temperature grid size (default is 100). If \code{NULL},
no annealing is performed.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL}, default hyperparameters will be used.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as it will be taken to be
\code{length(time_g)}.}

\item{tol_abs}{Tolerance on the absolute changes in the ELBO.}

\item{tol_rel}{Tolerance on the relative changes in the ELBO.}

\item{maxit}{Number of iterations for the mean-field variational Bayes algorithm. Default maxit = 1000.}

\item{n_cpus}{User-specified number of CPU to be used for parallel execution. Default n_cpus = 1 for serial execution.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run. Default is TRUE.}

\item{seed}{User-specified seed for reproducibility.}

\item{bool_scale}{Whether to standardise the variables. The per-variable
within-individual means are first computed, their mean and sd across
individuals are obtained, and used to scale the measurements. Default is TRUE.}

\item{bool_var_spec_prob}{Whether to use variable- (and factor-) specific
spike-and-slab probabilities. Default is TRUE.}

\item{show_factor_ppi_progress}{Whether to show a plot of the factor posterior
probabilities of inclusion as the algorithm progresses. Default is FALSE.}
}
\value{
An object containing the resulting estimates.
}
\description{
This function performs functional analysis to uncover shared latent dynamics
using Bayesian methods.
}
